### 2022-12-08 #4997 @danielskatz
👋 @PabRod - thanks for your submission. I'm the track editor, and will now start a scope query for the editors, since the number of lines of code is fairly small.  We should make a decision in the next week or two.

If we do proceed to review this submission, you will need to reduce the paper, since it's over 3 times as long as our guidance suggests. Much of the paper content now seems like it would better fit the documentation.  Please see https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
