### 2022-12-15 #4997 @danielskatz
👋 @PabRod - We'll go ahead and review this.  As I mentioned before, please reduce the paper, since it's over 3 times as long as our guidance suggests. Much of the paper content now seems like it would better fit the documentation. See https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain.

Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). After you make changes to your .md or .bib files, you can use the commands `@editorialbot check references` to check the references again, and `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

When you are done, please let me know, and we'll work on assigning an editor, then getting started
