### 2023-01-11 parallelwindfarms/byteparsing#13 @PabRod
Recycle draft paper 
## Problem

The [draft we submitted](https://github.com/parallelwindfarms/byteparsing/releases/tag/new-submission) to JOSS is [too long](https://github.com/openjournals/joss-reviews/issues/4997#issuecomment-1353318521). I'll be working on reducing its size, but it would be a pity if all the content gets lost.

## Solution

Luckily, we version-controlled the writing process. The first, exceedingly long submission is available via the [new-submission](https://github.com/parallelwindfarms/byteparsing/releases/tag/new-submission) tag, equivalent to commit 54515eb4823c64f3df2aaa99e1fec6b4cb8d1d17. Consider returning to it in the future, in case we want to extend the documentation.
