### 2023-01-11 c591d42
Minor edits
Added some urls and other minor editions
