### 2023-01-11 6dec2dd
Remove redundant examples
They are already available in the documentation
