+++ paper/paper.md
@@ -52,9 +52,9 @@ We were thus forced to write our own parser, with this list of requirements:
 - Performant enough, considering the use case where we have small ASCII headers and large contiguous blocks of floating point data.
 - Compliant with best practices such as automated unit testing and thorough documentation.
 
-# Architecture
+# A note on architecture
 
-Writing functional parser combinators is a staple of functional languages like Haskell or Ocaml [@frost_1989;@hutton_1992]. The paper "Monadic Parsing in Haskell" [@hutton_meijer_1998] gives a complete tutorial on how to write a basic recursive descent parser. Most of what Hutton and Meijer teach, carries over nicely to Python, once we take care of a few details. We've replaced some Haskell idioms by features that are considered more 'pythonic'.
+Writing functional parser combinators is a staple of functional languages like Haskell or Ocaml [@frost_1989;@hutton_1992]. The paper "Monadic Parsing in Haskell" [@hutton_meijer_1998] gives a complete tutorial on how to write a basic recursive descent parser. Most of what Hutton and Meijer teach, carries over nicely to Python once we take care of a few details. We've replaced some Haskell idioms by features that are considered more 'pythonic'. An extended description can be found in our documentation[^4].
 
 ## Functional parser combinators
 
@@ -254,130 +254,7 @@ assert parse_bytes(
     b'1hello') == "HELLO"
 ```
 
-## Examples of usage
-
-Although we provide a collection of useful parsers, the real power of this package is that it allows for parser composition.
-In this section we'll see a few examples about how this is done.
-
-### Parsing an e-mail address
-
-We begin by defining what characters are acceptable on an email:
-
-```python
-email_char = choice(
-                ascii_alpha_num, 
-                ascii_underscore, 
-                text_literal("."), 
-                text_literal("-")
-              )
-```
-
-Now, we abstract the information contained in an email as: `[user]@[host]`
-
-```python
-email = named_sequence(
-            user=some_char(email_char), # Step 1
-            _1=text_literal("@"), # Step 2
-            host=some_char(email_char) # Step 3
-        )
-```
-
-Now we are ready to parse:
-
-```python
-parse_bytes(email, b'p.rodriguez-sanchez@esciencecenter.nl')
-# returns {'user': b'p.rodriguez-sanchez', 'host': b'esciencecenter.nl'}
-```
-
-Notice that the parser created and populated a dictionary for us.
-
-### Parsing a list of e-mail addresses
-
-We can use parser composition to easily reuse what we learned in the previous section and apply it to a list of emails.
-Imagine the list of emails is stored as a newline-separated list:
-
-```python
-data = b"j.hidding@esciencecenter.nl\np.rodriguez-sanchez@esciencencenter.nl"
-```
-
-We'll want to identify and end-of-line character. Keep in mind that they may differ per OS:
-
-```python
-eol = choice(text_literal("\n\r"), text_literal("\n"))
-```
-
-Our parser is now generated by:
-
-```python
-list_of_emails = sep_by(email, eol)
-```
-
-This shows how we can slowly compose small and testable parsers to form larger more complicated ones. This composability and testability of each step is what make parser combinators such a powerful tool.
-
-### Binary example: PPM files
-In this example we show how we can mix ASCII and binary data. Here we parse Portable PixMap files (PPM). These files have a small ASCII header and the image itself in binary. The header looks something like this:
-
-```
-P6   # this marks the file type in the Netpbm family
-640 480
-255
-<<binary rgb values: 3*w*h bytes>>
-```
-
-This example has the following imports:
-
-```python
-import numpy as np
-from dataclasses import dataclass
[…]
