### 2023-01-11 5706aee
Move OpenFOAM example to docs
