### 2023-01-11 9bb3230
Move content about functional parsers from draft to documentation
Specifically, from ./paper/paper.md to ./docstr/functional-2.md

For now, we keep it out of the index
