### 2023-01-11 9733d72
Move content about parser grammar from draft to documentation
Specifically, from ./paper/paper.md to ./docstr/parser-grammar.md

For now, we keep it out of the index
