+++ paper/paper.md
@@ -57,137 +57,6 @@ We were thus forced to write our own parser, with this list of requirements:
 Writing functional parser combinators is a staple of functional languages like Haskell or Ocaml [@frost_1989;@hutton_1992]. The paper "Monadic Parsing in Haskell" [@hutton_meijer_1998] gives a complete tutorial on how to write a basic recursive descent parser. Most of what Hutton and Meijer teach, carries over nicely to Python once we take care of a few details. We've replaced some Haskell idioms by features that are considered more 'pythonic'.
 
 An extended description of the architecture can be found in our documentation[^4]. Those readers more interested in starting working right away will probably find our list of examples[^5] very handy.
-# Parser grammar
-
-## Primitives
-
-The boundary between what we consider *primitives* and derived parsers can become a bit vague, nevertheless here is a selection of the most important primitive parsers.
-
-`value(x)`
-: Always succeeds, doesn't consume input, returns `x`
-
-`fail(msg)`
-: Always fails, raises an exception with `msg` as text.
-
-`item`
-: Get a single byte from the stream.
-
-`text_literal(str)`
-: Succeeds if the next characters in the stream exactly match `str`.
-
-`char_pred(pred)`
-: Advances the end of the cursor if `pred` succeeds.
-
-`text_end_by(char)`
-: Advances the end of the cursor as until `char` is found.
-
-`push(x)`
-: Push a value on the auxiliary stack.
-
-`pop()`
-: Pop a value from the auxiliary stack.
-
-We also defined some derived parsers that should be useful in most contexts.
-
-`whitespace`
-: Matches tabs spaces and newlines.
-
-`eol`
-: Matches End of Line characters (_i.e.:_ either `\n` or `\n\r`).
-
-`integer`
-: Matches an integer value.
-
-`scientific_number`
-: Matches a floating point number, possibly in scientific notation.
-
-## Combinators
-
-<!-- This package is strongly based on Haskell's syntax and philosophy. But Python is obviously not Haskell. That is to say, there is no nice syntax for monadic actions. In order to solve this issue, we developed a similar grammar for Python. Below, we present a description of such a grammar. -->
-The next question is, how can we combine our primitive parsers? We already listed the main combinators briefly, here we go into a little more detail.
-
-`choice(*p)`
-: Tries every parser `p` in sequence until one succeeds. If all fail, `choice` gathers all exceptions and composes an error message from that.
-
-`sequence(*p)`
-: Runs every parser `p` in sequence and only returns the result of the last one.
-
-`named_sequence(**p)`
-: Runs every parser `p` in sequence and stores results in a dictionary. Keys that start with an underscore are not stored.
-
-`many(p)`
-: Runs the parser `p` until it fails. Returns a list of parsed items.
-
-`some(p)`
-: Parses `p` at least one time, or fail.
-
-The `many` and `some` combinators come in several flavours. Both have a variant called `many_char` and `some_char` that return a string instead of a list. One more flavour is `many_char_0` and `some_char_0` that do not flush the cursor.
-
-Some derived combinators help us shape a little language to describe grammars.
-
-`optional(p, default=None)`
-: Parses `p` or gives the default value.
-
-`tokenize(p)`
-: Parses `p` followed by optional whitespace. This makes sure we always start at the next token.
-
-`fmap(f)`
-: Takes a function `f`, returns a lambda that maps an argument through `f` to a `value` parser. That sounds complicated, but it allows us to pass a parsed result through `f` using the `>>` operator. For an example, see the PPM parser at the end of this paper.
-
-### `named_sequence` and `construct`
-
[…]
