### 2023-01-11 #4997 @PabRod
I've been working on:

1. Removing the excess of information from the draft.
2. Sending it to the project documentation.

Task 1 is basically done, and tomorrow I'll continue with task 2. This is still work in progress, so no action required.
