+++ setup.cfg
@@ -11,7 +11,7 @@ source = byteparsing
 
 [tool:pytest]
 testpaths = tests
-addopts = --cov --cov-report xml:cov.xml --cov-report term --cov-report html --mypy --flake8
+addopts = --cov --cov-report xml:cov.xml --cov-report term --cov-report html --mypy
 
 [mypy-numpy]
 ignore_missing_imports = True
