+++ setup.py
@@ -18,7 +18,6 @@ test_deps = [
     'pytest',
     'mypy',
     'pytest-cov',
-    'pytest-flake8',
     'pycodestyle',
     'pytest-mypy']
 
