### 2023-01-12 ff7e5ae
Move Cursors and Parsers sections to Architecture
