+++ docs/_sources/architecture.ipynb.txt
@@ -1,0 +1,444 @@
+{
+ "cells": [
+  {
+   "attachments": {},
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Architecture"
+   ]
+  },
+  {
+   "attachments": {},
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Cursors\n",
+    "\n",
+    "The most basic element of this package is the `Cursor`.\n",
+    "Roughly, a `Cursor` represents some data + a selection of it.\n",
+    "\n",
+    "Let's build our first `Cursor`:"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "from byteparsing import Cursor"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "Cursor(data=b'Hello world!', begin=0, end=3, encoding='utf-8')\n"
+     ]
+    }
+   ],
+   "source": [
+    "data = b\"Hello world!\"\n",
+    "\n",
+    "c = Cursor(data, begin=0, end=3)\n",
+    "print(c)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "The `Cursor` is implemented as a `dataclass`.\n",
+    "This means that it contains fields (particularly: `data`, `begin`, `end`, and `encoding`), and also methods.\n",
+    "\n",
+    "For instance, the method `content` returns the subsetted data (_i.e._: the `data` between `begin` and `end`):"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "b'Hel'"
+      ]
+     },
+     "execution_count": 3,
+     "metadata": {},
+     "output_type": "execute_result"
+    }
+   ],
+   "source": [
+    "c.content # This method is decorated as a property, so parentheses are not needed"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "The method `increment` returns a new cursor where `end` has been increased (by default, to `end + 1`)."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 4,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "Cursor(data=b'Hello world!', begin=0, end=3, encoding='utf-8')\n",
+      "Cursor(data=b'Hello world!', begin=0, end=4, encoding='utf-8')\n"
+     ]
+    }
+   ],
+   "source": [
+    "c = Cursor(data, begin=0, end=3)\n",
+    "print(c)\n",
+    "ci = c.increment()\n",
+    "print(ci)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "An interesting property of cursors is that they can be evaluated to a boolean. Particularly, a `Cursor` is `True` if and only if `end` is not at the end of the `data` string."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 5,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "cT = Cursor(data, begin=0, end=0)\n",
+    "cF = Cursor(data, begin=0, end=len(data))\n",
+    "\n",
[…]
