+++ docsrc/architecture.md
@@ -1,1 +1,0 @@
-../ARCHITECTURE
