### 2023-01-12 c0ec80d
Reorder: Parser grammar now appears after Architecture
