### 2023-01-12 parallelwindfarms/byteparsing#13 @PabRod
Solved from commit [eb15561](https://github.com/parallelwindfarms/byteparsing/commit/eb15561cee7b7078832168bfb01b75fc6ab1d7c1). Waiting to merge into master.
