### 2023-03-22 #5293 @editorialbot
[REVIEW]: byteparsing: a functional parser combinator for mixed ASCII/binary data
**Submitting author:** @PabRod (<a href="http://orcid.org/0000-0002-2855-940X">Pablo Rodríguez-Sánchez</a>)
**Repository:** https://github.com/parallelwindfarms/byteparsing
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.1.3
