### 2023-03-22 parallelwindfarms/byteparsing#15 @dvberkel
unable to install byteparsing via pip
As a reviewer of [_byteparsing: a functional parser combinator for mixed ASCII/Binary data_ ](https://github.com/openjournals/joss-reviews/issues/5293)
I tried to install `byteparsing` via pip
But was unable to do so

## What I did
I created a virtual environment via `python3 -m venv .env` and activiated it with

```fish
source .env/bin/activate.fish
```

Next a ran

```fish
pip install byteparsing
```

Which produces the following output

```plain
Collecting byteparsing
  Downloading byteparsing-0.1.0.tar.gz (13 kB)
     command: /home/dvberkel/scratch/foreign/byteparsing/.env/bin/python3 -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-4z81gl1b/byteparsing/setup.py'"'"'; __file__='"'"'/tmp/pip-install-4z81gl1b/byteparsing/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /tmp/pip-install-4z81gl1b/byteparsing/pip-egg-info
         cwd: /tmp/pip-install-4z81gl1b/byteparsing/
    Complete output (20 lines):
    WARNING: The wheel package is not available.
    WARNING: The wheel package is not available.
    WARNING: The wheel package is not available.
    WARNING: The wheel package is not available.
    WARNING: The wheel package is not available.
    WARNING: The wheel package is not available.
    WARNING: The wheel package is not available.
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-install-4z81gl1b/byteparsing/setup.py", line 25, in <module>
        setup(
      File "/home/dvberkel/scratch/foreign/byteparsing/.env/lib/python3.8/site-packages/setuptools/__init__.py", line 144, in setup
        _install_setup_requires(attrs)
      File "/home/dvberkel/scratch/foreign/byteparsing/.env/lib/python3.8/site-packages/setuptools/__init__.py", line 139, in _install_setup_requires
        dist.fetch_build_eggs(dist.setup_requires)
      File "/home/dvberkel/scratch/foreign/byteparsing/.env/lib/python3.8/site-packages/setuptools/dist.py", line 718, in fetch_build_eggs
        resolved_dists = pkg_resources.working_set.resolve(
      File "/home/dvberkel/scratch/foreign/byteparsing/.env/lib/python3.8/site-packages/pkg_resources/__init__.py", line 791, in resolve
        raise VersionConflict(dist, req).with_context(dependent_req)
    pkg_resources.ContextualVersionConflict: (docutils 0.19 (/tmp/pip-install-4z81gl1b/byteparsing/.eggs/docutils-0.19-py3.8.egg), Requirement.parse('docutils<0.19'), {'sphinx-rtd-theme'})
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
```
