### 2023-03-22 parallelwindfarms/byteparsing#15 @dvberkel
On an unrelated note, the pip command seems to install version `0.1.0` whereas your GitHub package list shows a version `0.1.2` to be available.
