### 2023-03-22 parallelwindfarms/byteparsing#16 @dvberkel
unable to install byteparser via git
As a reviewer of [_byteparsing: a functional parser combinator for mixed ASCII/Binary data_ ](https://github.com/openjournals/joss-reviews/issues/5293)
I tried to install `byteparsing` via git
But was unable to do so

## What I did
I cloned the repository via

```fish
 git clone https://github.com/parallelwindfarms/byteparsing.git
```
Entered the directory and created a virtual environment via `python3 -m venv .env` and activiated it.

Next I ran

```fish
pip install .
```

Which produces the following output

```plain
Processing /home/dvberkel/scratch/foreign/byteparsing/byteparsing
  Installing build dependencies ... done
    Preparing wheel metadata ... done
ERROR: Package 'byteparsing-0.1.2' requires a different Python: 3.8.10 not in '>=3.9,<3.11'
```
