### 2023-03-22 parallelwindfarms/byteparsing#16 @dvberkel
The problem is that my python version is `3.8.10`, which does not fall within the required range. Unfortunately this is not mentioned in the installation guide in the README.
