### 2023-03-22 dc04042
Solves #16
In the future, we should consider relaxing this requirement.
