### 2023-03-22 parallelwindfarms/byteparsing#15 @PabRod
Actually, v0.1.2 is [on PyPI](https://pypi.org/project/byteparsing/).

I manage to install it without a problem using conda:

```sh
conda create -n testenv python=3.9
conda activate testenv
pip install byteparsing
```
Returns:
```sh
  Downloading byteparsing-0.1.2-py3-none-any.whl (16 kB)
Collecting numpy<2.0.0,>=1.21.2
  Downloading numpy-1.24.2-cp39-cp39-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (17.3 MB)
Installing collected packages: numpy, byteparsing
Successfully installed byteparsing-0.1.2 numpy-1.24.2

```
