### 2023-03-23 parallelwindfarms/byteparsing#15 @PabRod
Any clue about what could be wrong, @jhidding?
