### 2023-03-26 parallelwindfarms/byteparsing#22 @inakleinbottle
JOSS Review issues and comments
Hi Folks,

First, I'm a big fan of parser combinators, I've used them several times to great effect (mostly using a Rust implementation but also in Haskell).

There are one or two issues that I think would help new users, both are about documentation (and more specifically, the readme):
 1. The description of what the package is for could really do with a bit more detail. You do cover this in the extended tutorial, but I think it would be worth repeating in the readme as a "first port of call".
 2. It might be good to include an example of the usage in the readme so people can get a very quick idea of how to use the library. I always find such examples very helpful when I'm looking for the right tool for the job. It doesn't need to be complicated.

Neither of these are blocking per-se, so if you have strong views then I have no problem completing the review but I do think they will be good for the package.

There is one other issue, which is not really part of my review. The inclusion of Numpy as a primary dependency is no strictly necessary, since the bulk of your library will work just fine without it. Consider making this an optional component for people who might not want a heavy install such as Numpy when it isn't necessary for their problem.
