### 2023-03-27 parallelwindfarms/byteparsing#22 @jhidding
Thanks! I will make these changes.
