### 2023-03-27 c5823e3
remove setup.py (pyproject.toml should suffice); update docs
