### 2023-03-27 parallelwindfarms/byteparsing#15 @jhidding
Yes, I think I upped the minimum Python version to 3.9 in between releases. So the newest one compatible with 3.8 is byteparsing 1.0. That one had a version conflict with docutils, which shouldn't have been in the dependencies in the first place.
