+++ pyproject.toml
@@ -33,6 +33,7 @@ Sphinx = "^5.3.0"
 nbsphinx = "^0.8.10"
 recommonmark = "^0.7.1"
 sphinx-rtd-theme = "^1.1.1"
+mypy = "^1.1.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
