### 2023-03-27 687b623
put array parser in separate module; rest of byteparsing should work without numpy now
