### 2023-03-27 64f4249
update api documentation to reflect different module layout
