+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "byteparsing"
-version = "0.1.2"
+version = "0.2.0"
 description = "Parser for mixed ASCII/binary data formats"
 authors = ["Johan Hidding <j.hidding@esciencecenter.nl>",
            "Pablo Rodríguez-Sánchez <p.rodriguez-sanchez@esciencecenter.nl>"]
