+++ README.md
@@ -6,12 +6,20 @@
 [![fair-software.eu](https://img.shields.io/badge/fair--software.eu-%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8B-orange)](https://fair-software.eu)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4734194.svg)](https://doi.org/10.5281/zenodo.4734194)
 
-Parser for mixed ASCII/binary data. Features:
+Byteparsing is a package for **mixed text and binary** parsing in Python. The main driver for developing this package was to write a parser for binary OpenFOAM files. The binary file format in OpenFOAM is "special". It is the same as the ASCII based text format, except where large blocks of floating point data are concerned.
 
-- Works extremely well with memory-mapped Numpy arrays
-- Included parsers:
-    - OpenFOAM
-    - Examples: PLY, PPM
+When **not to use** `byteparsing`:
+
+* You just need to parse some text: use `pyparsing`, it is the industry's standard.
+
+**Do use** `byteparsing` if:
+
+* You need to tinker with large binary OpenFOAM files directly from Python.
+* There is a different package that does not adhere to data standards and hacked together its own mixed ASCII/binary file format. You will have to roll out your own parser. Byteparsing can make this easier.
+
+**Coolest feature**:
+
+* Works with `mmap` and `numpy`! This means you can open the file without reading it entirely into memory, change the NumPy array data and the changes are automatically saved to disk.
 
 ## Documentation
 Our [documentation](https://parallelwindfarms.github.io/byteparsing) explains the architecture behind `byteparsing` and shows some examples of parsing mixed binary and ASCII data.
