### 2023-03-27 parallelwindfarms/byteparsing#22 @jhidding
Updated the README with better intro text from the documentation, the PPM example (minimal and shows mixing of ascii and binary).

Refactored the `array` parser into its own submodule, skip tests that need numpy if import fails.
