### 2023-04-11 #5293 @AoifeHughes
@Pabrod this all looks great to me, whilst I'm doing some final checks could you make sure that:
- [ ] There has been a tagged release made on GitHub (and let me know the version number if it's changed)
- [ ] Make an archive on either https://zenodo.org/ https://figshare.com/ (or another site if those aren't suitable). Please update me with the DOI and link once you have those

With both of these done, we should be able to make some final checks for publication once I've verified everything. If you've any questions please ping me!
