### 2023-04-18 #5293 @PabRod
Dear @AoifeHughes,

I've just updated the [release](https://github.com/parallelwindfarms/byteparsing/releases/tag/0.1.3) and [Zenodo archive](https://zenodo.org/record/7839894#.ZD5RjtJByLs).

I'm just back from my holidays, thus the delay in my answer :)
