### 2023-04-20 #5293 @danielskatz
Hi all - I'll proofread this now.

FYI, @AoifeHughes - we need the code archive to be a DOI, so I'll fix that.

@PabRod - We also want the archive metadata to match the metadata of the paper. Please change the title in the zenodo metadata to match that of the paper.  (See the section labeled "How can I edit the metadata of a published record?" in https://help.zenodo.org for how to do this.)
