### 2023-04-20 #5293 @danielskatz
@PabRod - in addition to the metadata change I requested above, I've also found a few small issues in the paper, and created https://github.com/parallelwindfarms/byteparsing/pull/23 to fix them. Please also merge this, or let me know what you disagree with.  Once these two things are done, we can move to acceptance and publication.
