### 2023-04-21 #5293 @PabRod
@jhidding, this will require an action from your side. You are the owner of the relevant Zenodo archive and I don't have edit permissions.

> @PabRod - We also want the archive metadata to match the metadata of the paper. Please change the title in the zenodo metadata to match that of the paper. (See the section labeled "How can I edit the metadata of a published record?" in https://help.zenodo.org for how to do this.)
