### 2023-04-21 #5293 @danielskatz
@PabRod and @jhidding - this proof looks good to me, so once [the Zenodo title is fixed](https://github.com/openjournals/joss-reviews/issues/5293#issuecomment-1517620059), we can accept and publish this submission.
